Refactor app to unclutter App.js Add a blog-list component which renders the blogs retrievied from the server
